libsyn123: silence a warning about -unsigned (bug 317).
git-svn-id: svn://scm.orgis.org/mpg123/trunk@4960 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libsyn123/resample.c
@@ -2005,7 +2005,7 @@ syn123_resample_intotal_64(long inrate, long outrate, int64_t outs)
 	// First offset is -inrate.
 	// You may want to work it out for yourself. Or trust me;-)
 	int err;
-	uint64_t vtot = muloffdiv64(outs, vinrate, -vinrate, voutrate, &err, NULL);
+	uint64_t vtot = muloffdiv64(outs, vinrate, -(int64_t)vinrate, voutrate, &err, NULL);
 	if(err)
 		return SYN123_OVERFLOW;
 	if(vtot == UINT64_MAX)
